:truck: Refactor AIG to NetworkX conversion into a separate module
